Move collection views to own file
